requirements.txt
# ============================================================
# Data Engineering Portfolio — Projects 6-10
# Python 3.10+
# ============================================================

# Kafka
kafka-python==2.0.2

# PySpark + Delta Lake
pyspark==3.5.0
delta-spark==3.1.0

# Apache Flink (PyFlink)
apache-flink==1.18.1

# Airflow
apache-airflow==2.8.1
apache-airflow-providers-google==10.13.1
apache-airflow-providers-amazon==8.19.0
apache-airflow-providers-snowflake==5.3.1
apache-airflow-providers-elasticsearch==5.3.1

# Cloud — BigQuery
google-cloud-bigquery==3.17.2
google-cloud-storage==2.16.0

# Cloud — AWS S3
boto3==1.34.34
s3fs==2024.2.0

# Cloud — Snowflake
snowflake-connector-python==3.7.1

# NoSQL — Cassandra (Project 10)
cassandra-driver==3.29.1

# Elasticsearch (Project 8)
elasticsearch==8.12.1

# ClickHouse (Project 7)
clickhouse-connect==0.7.6

# DuckDB (Project 9)
duckdb==0.10.1

# SSE Client (Project 10 — Wikimedia)
sseclient-py==1.8.0

# Data Processing
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0

# HTTP
requests==2.31.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
